+++ jdaviz/tests/test_user_api.py
@@ -1,4 +1,6 @@
 from jdaviz.configs.imviz.tests.utils import BaseImviz_WCS_WCS
+import pytest
+import re
 
 
 # This applies to all viz but testing with Imviz should be enough.
@@ -47,3 +49,79 @@ def test_toggle_api_hints(specviz_helper):
     assert specviz_helper.app.state.show_api_hints is True
     specviz_helper.toggle_api_hints()
     assert specviz_helper.app.state.show_api_hints is False
+
+
+def test_wildcard_matching_selected(imviz_helper, multi_extension_image_hdu_wcs):
+    default_choices = ['1: [SCI,1]',
+                       '2: [MASK,1]',
+                       '3: [ERR,1]',
+                       '4: [DQ,1]']
+
+    # Testing directly
+    ldr = imviz_helper.loaders['object']
+    ldr.object = multi_extension_image_hdu_wcs
+    extension_obj = ldr.importer.extension
+    # Default selection
+    assert extension_obj.selected == [default_choices[0]]
+    # Resetting to []
+    # Note this can't be done by setting selected = [], is this intentional?
+    extension_obj.selected.pop(0)
+
+    assert extension_obj.choices == default_choices
+    extension_obj.multiselect = False
+
+    err_str1 = "not all items in"
+    err_str2 = f"are one of {extension_obj.choices}, reverting selection to []"
+    with pytest.raises(ValueError,
+                       match=re.escape(f"{err_str1} ['bad *'] {err_str2}")):
+        ldr.importer._obj.user_api.extension = 'bad *'
+
+    with pytest.raises(ValueError,
+                       match=re.escape(f"{err_str1} ['bad *', '* result'] {err_str2}")):
+        ldr.importer._obj.user_api.extension = ['bad *', '* result']
+
+    with pytest.raises(ValueError,
+                       match=re.escape(f"{err_str1} ['another', 'bad * result'] {err_str2}")):
+        ldr.importer._obj.user_api.extension = ['another', 'bad * result']
+
+    # Check that selected is still/reverted successfully to []
+    assert extension_obj.selected == []
+
+    test_selections = {
+        # Test all
+        '*': extension_obj.choices,
+        # Test repeats
+        ('*', '*:*'): extension_obj.choices,
+        # Test single selection
+        '1:*': [extension_obj.choices[0]],
+        # Test multi-wildcard
+        '*S*': extension_obj.choices[:2],
+        # Test multi-selection
+        ('*ERR*', '*DQ*'): extension_obj.choices[2:]}
+
+    for selection, expected in test_selections.items():
+        ldr.importer._obj.user_api.extension = selection
+        assert extension_obj.multiselect is True
+        assert extension_obj.selected == expected
+        # Reset
+        extension_obj.selected = []
+        extension_obj.multiselect = False
+
+
+@pytest.mark.parametrize(
+    ("selection", "matches"), [
+        ('*', (0, 1, 2, 3)),
+        (('*', '*:*'), (0, 1, 2, 3)),
+        ('1:*', (0,)),
+        ('*S*', (0, 1)),
+        (('*ERR*', '*DQ*'), (2, 3))])
+def test_wildcard_matching_through_load(imviz_helper, multi_extension_image_hdu_wcs,
+                                        selection, matches):
+    data_labels = ['Image[SCI,1]',
+                   'Image[MASK,1]',
+                   'Image[ERR,1]',
+                   'Image[DQ,1]']
+
+    # Through load
+    imviz_helper.load(multi_extension_image_hdu_wcs, extension=selection)
+    assert imviz_helper.data_labels == [data_labels[i] for i in matches]